:recycle: refactor: move landing page to app router
