+++ frontend/src/app/_components/LandingHero/index.tsx
@@ -1,14 +1,14 @@
-import { Box, Button, Container, Typography, Unstable_Grid2 as Grid } from "@mui/material";
-import { useTheme } from "@mui/material/styles";
+"use client";
+
+import { Box, Button, Container, Unstable_Grid2 as Grid, Typography } from "@mui/material";
 import Image from "next/image";
 
-import { Link } from "@/components";
+import { Link } from "@/app/components/Link";
 import Hero from "~/public/static/landing/hero.webp";
 
 import { OrganizationsSlider } from "./OrganizationsSlider";
 
 export const LandingHero: React.FC = () => {
-  const theme = useTheme();
   return (
     <>
       <Box
@@ -92,7 +92,6 @@ export const LandingHero: React.FC = () => {
             alt=""
             priority
             sizes={`
-              ${theme.breakpoints.down("sm")} 0vw,
               40vw,
             `}
           />
