+++ frontend/src/app/_components/LandingListings.tsx
@@ -1,10 +1,12 @@
+"use client";
+
 import { Box, Button, Container, Grid, Typography } from "@mui/material";
 import { useTheme } from "@mui/material/styles";
 import Image from "next/image";
 
 import Social from "~/public/img/gang.jpg";
 
-import { NextLinkComposed } from "../../Link";
+import { NextLinkComposed } from "@/app/components/Link";
 
 export const LandingListings: React.FC = () => {
   const theme = useTheme();
