+++ frontend/src/app/_components/LandingPromo.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { Container, Grid } from "@mui/material";
 import { styled } from "@mui/material/styles";
