+++ frontend/src/app/layout.tsx
@@ -14,7 +14,7 @@ export const metadata: Metadata = {
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
-    <html lang="en">
+    <html lang="en" data-color-scheme="light" suppressHydrationWarning>
       <body>
         <ThemeRegistry>
           <ApolloWrapper>
